Move logic to find scenario by parsing a single cell's content to the test context (where basic scenario finding was already)
